requirements.txt
numpy
scipy
streamlit
plotly
pandas
mysql-connector-python
